requirements.txt
frappe>=14.0.0
requests>=2.28.0
python-socketio>=5.7.0
eventlet>=0.33.0
PyJWT>=2.4.0
anthropic>=0.3.0
openai>=0.27.0
redis>=4.3.0
celery>=5.2.0